+++ src/comment_header.rs
@@ -8,6 +8,7 @@ use crate::Error;
 
 const COMMENT_MAGIC: &[u8] = &[0x4f, 0x70, 0x75, 0x73, 0x54, 0x61, 0x67, 0x73];
 
+/// Allows querying and modification of an Opus comment header
 #[derive(Derivative)]
 #[derivative(Debug)]
 pub struct CommentHeader<'a> {
@@ -26,12 +27,21 @@ impl<'a> CommentHeader<'a> {
         reader.read_exact(data).map_err(|_| Error::MalformedCommentHeader)
     }
 
+    /// Constructs an empty `CommentHeader`. The comment data will be placed in
+    /// the supplied `Vec`. Any existing content will be discarded.
     pub fn empty(data: &'a mut Vec<u8>) -> CommentHeader<'a> {
         CommentHeader { data, vendor: String::new(), user_comments: Vec::new() }
     }
 
+    /// Sets the vendor field.
     pub fn set_vendor(&mut self, vendor: &str) { self.vendor = vendor.to_string(); }
 
+    /// Attempts to parse the supplied `Vec` as an Opus comment header. An error
+    /// is returned if the header is believed to be corrupt, otherwise an
+    /// `Option` is returned containing either the parsed header or `None`
+    /// if the comment magic string was not found. This enables
+    /// distinguishing between a corrupted comment header and a packet which
+    /// does not appear to be a comment header.
     pub fn try_parse(data: &'a mut Vec<u8>) -> Result<Option<CommentHeader<'a>>, Error> {
         let identical = data.iter().take(COMMENT_MAGIC.len()).eq(COMMENT_MAGIC.iter());
         if !identical {
@@ -57,6 +67,7 @@ impl<'a> CommentHeader<'a> {
         Ok(Some(result))
     }
 
+    /// Returns the first mapped value for the specified key.
     pub fn get_first(&self, key: &str) -> Option<&str> {
         for (k, v) in self.user_comments.iter() {
             if k == key {
@@ -66,17 +77,42 @@ impl<'a> CommentHeader<'a> {
         None
     }
 
+    /// Removes all mappings for the specified key.
     pub fn remove_all(&mut self, key: &str) { self.user_comments.retain(|(k, _)| key != k); }
 
+    /// If the key already exists, update the first mapping's value to the one
+    /// supplied and discard any later mappings. If the key does not exist,
+    /// append the mapping to the end of the list.
     pub fn replace(&mut self, key: &str, value: &str) {
-        self.remove_all(key);
-        self.append(key, value);
+        let mut found = false;
+        self.user_comments.retain_mut(|(k, ref mut v)| {
+            if k == key {
+                if found {
+                    // If we have already found the key, discard this mapping
+                    false
+                } else {
+                    *v = value.to_string();
+                    found = true;
+                    true
+                }
+            } else {
+                true
+            }
+        });
+
+        // If the key did not exist, we append
+        if !found {
+            self.append(key, value);
+        }
     }
 
+    /// Appends the specified mapping.
     pub fn append(&mut self, key: &str, value: &str) {
         self.user_comments.push((String::from(key), String::from(value)));
     }
 
+    /// Attempts to parse the first mapping for the specified key as the
+    /// fixed-point Decibel representation used in Opus comment headers.
     pub fn get_gain_from_tag(&self, tag: &str) -> Result<Option<FixedPointGain>, Error> {
         let parsed =
             self.get_first(tag).map(|v| v.parse::<FixedPointGain>().map_err(|_| Error::InvalidR128Tag(v.to_string())));
@@ -87,6 +123,7 @@ impl<'a> CommentHeader<'a> {
         }
     }
 
+    /// Returns the album gain if present, else the track gain, else `None`.
     pub fn get_album_or_track_gain(&self) -> Result<Option<FixedPointGain>, Error> {
         for tag in [TAG_ALBUM_GAIN, TAG_TRACK_GAIN].iter() {
             if let Some(gain) = self.get_gain_from_tag(tag)? {
@@ -96,6 +133,9 @@ impl<'a> CommentHeader<'a> {
         Ok(None)
     }
 
+    /// Applies the specified delta to either or both of the album and track
+    /// gains if present. If neither as present, this function will do
+    /// nothing.
     pub fn adjust_gains(&mut self, adjustment: FixedPointGain) -> Result<(), Error> {
         if adjustment.is_zero() {
             return Ok(());
@@ -109,7 +149,7 @@ impl<'a> CommentHeader<'a> {
         Ok(())
     }
 
-    pub fn commit(&mut self) {
+    fn commit(&mut self) {
         //TODO: Look more into why we can't use https://github.com/rust-lang/rust/pull/46830
         let mut writer = Cursor::new(Vec::new());
         writer.write_all(COMMENT_MAGIC).unwrap();
@@ -219,5 +259,51 @@ mod tests {
             Err(Error::MalformedCommentHeader) => {}
             _ => assert!(false, "Wrong error for malformed header"),
         };
+    }
+
+    #[test]
+    fn replace_appends_on_missing() {
+        let key = "foo";
+        let value = "bar";
+
+        let mut data_1 = Vec::new();
+        let mut header_1 = CommentHeader::empty(&mut data_1);
+        header_1.append("v0", "k0");
+        header_1.append(key, value);
+        header_1.append("v1", "k1");
+
+        let mut data_2 = Vec::new();
+        let mut header_2 = CommentHeader::empty(&mut data_2);
+        header_2.append("v0", "k0");
+        header_2.replace(key, value);
+        header_2.append("v1", "k1");
+
+        assert_eq!(header_1, header_2);
+    }
+
+    #[test]
+    fn replace_replaces_on_duplicates() {
+        let mut data_1 = Vec::new();
+        let mut header_1 = CommentHeader::empty(&mut data_1);
+        header_1.append("v0", "k0");
+        header_1.append("v1", "k1");
+        header_1.append("v2", "k2");
+        header_1.append("v3", "k3");
+        header_1.append("v2", "k4");
+        header_1.append("v5", "k5");
+        header_1.append("v2", "k6");
+        header_1.append("v7", "k7");
+        header_1.replace("v2", "k8");
+
+        let mut data_2 = Vec::new();
+        let mut header_2 = CommentHeader::empty(&mut data_2);
+        header_2.append("v0", "k0");
+        header_2.append("v1", "k1");
+        header_2.append("v2", "k8");
+        header_2.append("v3", "k3");
+        header_2.append("v5", "k5");
+        header_2.append("v7", "k7");
+
+        assert_eq!(header_1, header_2);
     }
 }